requirements.txt
asgiref==3.5.2
Django==4.0.6
django-email-users==1.1.0
django-filter==22.1
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.1
signals==0.0.2
sqlparse==0.4.2
whitenoise==6.2.0
